refactored controller, stream lifecycle
